Update intune-deviceconfig-devicecompliancepolicyassignment-create.md
The correct verb is "assign".  When calling "assignments" as in original doc, the error is displayed as below:

     | {"error":{"code":"No method match route template","message":"No OData route exists that match template ~/singleton/navigation/key/navigation with http verb POST for request
     | /DeviceConfiguration_2108/StatelessDeviceConfigurationFEService/deviceManagement/deviceCompliancePolicies('ce0ac224-9f0e-4988-a2dd-78b474b72e91')/assignments.","innerError":{"date":"2021-09-02T21:52:54","request-id":"11b18ba2-83a3-49e1-aa50-fb8a7645878d","client-request-id":"11b18ba2-83a3-49e1-aa50-fb8a7645878d"}}}

"assign" is the path at the end of request, which can be seen by using F12 Developer Tools in MEM portal, when creating an assignment, and "assign" verb works fine. 

Suggestion for documentation improvement: Add more examples for group-based licensing, as this involves a different payload (groupAssignmentTarget inherits from deviceAndAppManagementAssignmentTarget , but that's not immediately obvious)